public index.php
In this file, instead of travel through all the properties of $_SERVER, just shows the relevant ones.
Since there are sensible information that might help any hacker to find the end-server where the final site will be
hosted.

In the meanwhile, we are still displaying relevant information that will help us continue during the early steps of the
development of the framework.