Controllers and Routing
In this Commit

Created the base controller (Controller.php) and an impplementation of it, home controller.
These files orchestrate how to work a view.

Route will receive a raw request and follow it's content all the way until a controller has been selected and the appropriate method
being called.
RouteType is an enum that will help the Route class to find the most appropriate controller.

Added files to the resources.
header is the segment for the layout that will host the first(header) section of the layout.
layout, the whole structure for a page to be loaded.
about & home. these files will create the content of the "main" section of the page, later being injected into layout.

Application. implement the Enroute method from Route to begin the process.